+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Added
 
-- Add marker top position to draw event
+- Add marker top and center position to draw event
 
 ## [0.14.0] - 2024-10-03